flow_configuration_repository_test_suite: rename file
